refactor: result 패키지 생성
