+++ src/main/java/blackjack/domain/BlackjackGame.java
@@ -4,6 +4,8 @@ import blackjack.domain.card.Card;
 import blackjack.domain.card.Deck;
 import blackjack.domain.card.HoldCards;
 import blackjack.domain.entry.*;
+import blackjack.domain.result.BettingResult;
+import blackjack.domain.result.GameResult;
 
 import java.util.List;
 import java.util.Map;
