+++ src/main/java/blackjack/domain/result/PlayerOutcome.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.result;
 
 import static blackjack.domain.BlackjackGame.BLACKJACK_NUMBER;
